+++ src/math/integration/mod.rs
@@ -1,0 +1,9 @@
+//!
+//! # Integration routines
+//!
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+pub mod d1;
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
